Export source material insertion types from schema
